docs: update code command help
This file has all revisions through January 3, 2022, for review by the Hammer team. Question: For commands with the sub-command test, should the test command be explained the same as as the snyk test command: test a project for open source vulnerabilities and license issues? At this time, the code and iac commands say: test for any known issue. The container command says: test for any known vulnerabilities.